model/requirements.txt
flask>=2.0.0
flask-cors>=4.0.0
requests==2.31.0
numpy>=1.24.0
torch>=2.0.0
transformers>=4.31.0
pyannote.audio>=3.1.0
huggingface_hub==0.17.3
python-dotenv>=1.0.0
pydub==0.25.1
scipy==1.11.2
librosa==0.10.1
ffmpeg-python>=0.2.0
sentencepiece>=0.1.99  # BART model için gerekli
sacremoses>=0.0.53     # BART tokenizer için gerekli
protobuf>=3.20.0       # Model serileştirme için gerekli
# GPU dependencies
# NOT: PyTorch CUDA için pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118 komutunu kullanın
accelerate>=0.21.0  # Hugging Face modelleri için GPU hızlandırma 